Reorder khash headers
